requirements.txt
requests>=2.25.0
Pillow>=8.0.0
torch>=1.9.0
numpy>=1.21.0
minio>=7.0.0